feat: Enhance event creation and editing with date range handling
- Update CreateEventDialog to clear time and recurring settings when an end date is set.
- Disable time and recurring options in CreateEventDialog when an end date is selected.
- Modify EditEventDialog to handle date ranges and clear time and recurring settings when an end date is set.
- Introduce EventCard component to encapsulate event display and editing logic.
- Refactor event filtering logic and utility functions for better maintainability.
- Ensure TypeSelector is now a required field in the event creation process.
